tests/wpt/web-platform-tests/offscreen-canvas/the-canvas-state/2d.state.saverestore.strokeStyle.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.state.saverestore.strokeStyle</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.state.saverestore.strokeStyle</h1>
<p class="desc">save()/restore() works for strokeStyle</p>


<script>
var t = async_test("save()/restore() works for strokeStyle");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

// Test that restore() undoes any modifications
var old = ctx.strokeStyle;
ctx.save();
ctx.strokeStyle = "#ff0000";
ctx.restore();
_assertSame(ctx.strokeStyle, old, "ctx.strokeStyle", "old");

// Also test that save() doesn't modify the values
ctx.strokeStyle = "#ff0000";
old = ctx.strokeStyle;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "#ff0000"
ctx.save();
_assertSame(ctx.strokeStyle, old, "ctx.strokeStyle", "old");
ctx.restore();

t.done();

});
</script>
